Moved about, simplifed Theme colour settings and moved Font Size to it
