+++ lib/settings/_general.dart
@@ -60,22 +60,6 @@ class SettingsGeneralFragment extends StatelessWidget {
         ]);
   }
 
-  void _createTweetFontSizeDialog(BuildContext context) async {
-    int? selectedFontSize = await showDialog<int>(
-      context: context,
-      builder: (context) => FontSizePickerDialog(initialFontSize: _getOptionTweetFontSizeValue(context)),
-    );
-    if (selectedFontSize != null) {
-      PrefService.of(context).set<int>(optionTweetFontSize, selectedFontSize);
-    }
-  }
-
-  int _getOptionTweetFontSizeValue(BuildContext context) {
-    int optionTweetFontSizeValue =
-        PrefService.of(context).get<int>(optionTweetFontSize) ?? DefaultTextStyle.of(context).style.fontSize!.round();
-    return optionTweetFontSizeValue;
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -202,12 +186,6 @@ class SettingsGeneralFragment extends StatelessWidget {
             subtitle: Text(L10n.of(context).leaner_feeds_description),
             pref: optionLeanerFeeds,
           ),
-          PrefButton(
-            title: Text(L10n.of(context).tweet_font_size_label),
-            subtitle: Text(L10n.of(context).tweet_font_size_description),
-            onTap: () => _createTweetFontSizeDialog(context),
-            child: Text('${_getOptionTweetFontSizeValue(context)} px'),
-          ),
         ]),
       ),
     );
@@ -248,8 +226,8 @@ class DownloadTypeSettingState extends State<DownloadTypeSetting> {
               DeviceInfoPlugin plugin = DeviceInfoPlugin();
               AndroidDeviceInfo android = await plugin.androidInfo;
               var storagePermission = android.version.sdkInt < 30
-                ? await Permission.storage.request()
-                : await Permission.manageExternalStorage.request();
+                  ? await Permission.storage.request()
+                  : await Permission.manageExternalStorage.request();
               if (storagePermission.isGranted) {
                 String? directoryPath = await FilePicker.platform.getDirectoryPath();
                 if (directoryPath == null) {
@@ -277,66 +255,6 @@ class DownloadTypeSettingState extends State<DownloadTypeSetting> {
             ),
             child: Text(L10n.current.choose),
           )
-      ],
-    );
-  }
-}
-
-class FontSizePickerDialog extends StatefulWidget {
-  /// initial selection for the slider
-  final int initialFontSize;
-
-  const FontSizePickerDialog({Key? key, required this.initialFontSize}) : super(key: key);
-
-  @override
-  FontSizePickerDialogState createState() => FontSizePickerDialogState();
-}
-
-class FontSizePickerDialogState extends State<FontSizePickerDialog> {
-  /// current selection of the slider
-  late int tweetFontSize;
-
-  @override
-  void initState() {
-    super.initState();
-    tweetFontSize = widget.initialFontSize;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    double defaultFontSize = DefaultTextStyle.of(context).style.fontSize!;
-    double minFontSize = defaultFontSize - 4;
-    double maxFontSize = defaultFontSize + 8;
-    return AlertDialog(
-      title: Text(L10n.of(context).tweet_font_size_label),
-      content: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text('$tweetFontSize px'),
-          Slider(
-            value: tweetFontSize.toDouble(),
-            min: minFontSize,
-            max: maxFontSize,
-            divisions: ((maxFontSize - minFontSize) / 2).round(),
-            label: '$tweetFontSize px',
-            onChanged: (value) {
-              setState(() {
-                tweetFontSize = value.round();
-              });
-            },
-          ),
-        ]
-      ),
-
-      actions: [
-        TextButton(onPressed: () => Navigator.pop(context), child: Text(L10n.of(context).cancel)),
-        TextButton(
-            onPressed: () async {
-              Navigator.pop(context, tweetFontSize);
-            },
-            child: Text(L10n.of(context).save)
-        )
       ],
     );
   }
