+++ lib/settings/_theme.dart
@@ -1,3 +1,5 @@
+import 'dart:ffi';
+
 import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:flutter/material.dart';
 import 'package:squawker/constants.dart';
@@ -7,6 +9,22 @@ import 'package:pref/pref.dart';
 
 class SettingsThemeFragment extends StatelessWidget {
   const SettingsThemeFragment({Key? key}) : super(key: key);
+
+  int _getOptionTweetFontSizeValue(BuildContext context) {
+    int optionTweetFontSizeValue =
+        PrefService.of(context).get<int>(optionTweetFontSize) ?? DefaultTextStyle.of(context).style.fontSize!.round();
+    return optionTweetFontSizeValue;
+  }
+
+  void _createTweetFontSizeDialog(BuildContext context) async {
+    int? selectedFontSize = await showDialog<int>(
+      context: context,
+      builder: (context) => FontSizePickerDialog(initialFontSize: _getOptionTweetFontSizeValue(context)),
+    );
+    if (selectedFontSize != null) {
+      PrefService.of(context).set<int>(optionTweetFontSize, selectedFontSize);
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -34,9 +52,11 @@ class SettingsThemeFragment extends StatelessWidget {
               fullWidth: false,
               pref: optionThemeColorScheme,
               items: FlexScheme.values
-                  .getRange(0, FlexScheme.values.length - 1)
-                  .sorted((a, b) => a.name.compareTo(b.name))
-                  .map((scheme) => DropdownMenuItem(value: scheme.name, child: Text(scheme.name.capitalize)))
+                  .where((scheme) => scheme.name.endsWith("M3"))
+                  .map((scheme) => DropdownMenuItem(
+                        value: scheme.name,
+                        child: Text(scheme.name.substring(0, scheme.name.length - 2).capitalize),
+                      ))
                   .toList()),
           PrefSwitch(
             title: Text(L10n.of(context).true_black),
@@ -45,8 +65,69 @@ class SettingsThemeFragment extends StatelessWidget {
               L10n.of(context).use_true_black_for_the_dark_mode_theme,
             ),
           ),
+          PrefButton(
+            title: Text(L10n.of(context).tweet_font_size_label),
+            subtitle: Text(L10n.of(context).tweet_font_size_description),
+            onTap: () => _createTweetFontSizeDialog(context),
+            child: Text('${_getOptionTweetFontSizeValue(context)} px'),
+          ),
         ]),
       ),
+    );
+  }
+}
+
+class FontSizePickerDialog extends StatefulWidget {
+  /// initial selection for the slider
+  final int initialFontSize;
+
+  const FontSizePickerDialog({Key? key, required this.initialFontSize}) : super(key: key);
+
+  @override
+  FontSizePickerDialogState createState() => FontSizePickerDialogState();
+}
+
+class FontSizePickerDialogState extends State<FontSizePickerDialog> {
+  /// current selection of the slider
+  late int tweetFontSize;
+
+  @override
+  void initState() {
+    super.initState();
+    tweetFontSize = widget.initialFontSize;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    double defaultFontSize = DefaultTextStyle.of(context).style.fontSize!;
+    double minFontSize = defaultFontSize - 4;
+    double maxFontSize = defaultFontSize + 8;
+
+    return AlertDialog(
+      title: Text(L10n.of(context).tweet_font_size_label),
+      content: Column(mainAxisAlignment: MainAxisAlignment.center, mainAxisSize: MainAxisSize.min, children: [
+        Text('$tweetFontSize px'),
+        Slider(
+          value: tweetFontSize.toDouble(),
+          min: minFontSize,
+          max: maxFontSize,
+          divisions: ((maxFontSize - minFontSize) / 2).round(),
+          label: '$tweetFontSize px',
+          onChanged: (value) {
+            setState(() {
+              tweetFontSize = value.round();
+            });
+          },
+        ),
+      ]),
+      actions: [
+        TextButton(onPressed: () => Navigator.pop(context), child: Text(L10n.of(context).cancel)),
+        TextButton(
+            onPressed: () async {
+              Navigator.pop(context, tweetFontSize);
+            },
+            child: Text(L10n.of(context).save))
+      ],
     );
   }
 }
